Bug 5561 - This bug revealed many other bugs. The tools menu in the overview header needed to be restructured. The reports in the performance areas all had their own individual problems. Seems that the reporting in the performance area was never complete.

+++ dao/src/test/java/com/inthinc/pro/dao/util/MetricConversionUtilTest.java
@@ -51,7 +51,7 @@ public class MetricConversionUtilTest
     {
         Long expecedcKPL = 10L;
         Long actualMPG = 23L;
-        Long kilometersPerLiter = MeasurementConversionUtil.fromMPGtoKPL(actualMPG);
+        Long kilometersPerLiter = MeasurementConversionUtil.fromMPGtoKPL(actualMPG).longValue();
         Assert.assertEquals(expecedcKPL, kilometersPerLiter);
     }
